refactor(filter): Pull out filter mod
